start going back to the old intonation and timing implementation
This uses intonation-, timing- and helper-code from the newly surfaced
code drop (in this case from tools/uni_port, as this is the oldest
"modern" code in the archives). This still needs to be compared with
code from microdectalk (which is much older) to get back to the level
of 4.2/4.3.