implement software volume-control [:vs nn] based on dtexp-code
Description from EAB:

Since the system is tuned "hot" the volume is at a max and these
controls allow it to be reduced. It works by reducing the the gain of
the final cascade resonator (different depending on what version of the
vocal tract is used), and the two other noise source gains GH and GF.
You simply need do decrease these all by the same db amount and the
results tract with good fidelity.