fix floating point voice tract model
this fixes the floating point vtm to work with the "old" synthesizer
again. Also allow tuning of the floating point output gain. For this
all voice definitions were re-dumped and the dumper modified not to
emit #ifndefs for FP_VTM.
Remove the old floating-point VTM imported from earlier sources with
wrong gains again.