vtmtable: add calculation for radius/cosine-tables
This adds a direct calculation for the values in the radius-
and cosine-tables used my the integer vtm. This could be used
to replace the table-lookup with a higher precision calculation
(the lokup must not be divided by 8 like it is done now) and
remove the zapping in vtmfunc.h (as a memory out-of-bounds is
no longer possible). Adds a significant computational overhead,
so not used for now.