FP VTM: A few more 22kHz changes
Not much better.
There is the inherent problem, that the connection between the
phoneme-module and the VTM is "just" 16 bit signed values. These can
overflow when using 22kHz sample rate causing many weird problems.
This currently papers over this bug by checking for negative values
on variables which are used as indices but this is not a solution.

The correct solution would be to make the connection 32bit signed
between these two modules, but this means major changes to both
modules which most likely breaks stuff relying on 16bit signed
behaviour.

For this reason, 22kHz is probably not achievable in DECtalk.