vtm: tune limit-cycle rampdown time
the limit-cycle rampdown-code ramped audio down abruptly after the last
audio was generated, giving the filters no time to settle and cutting of
sharp sounds at the end, like "t" in "street".
As the code was only meant to squelch the limit-cycles of the filters in
idle, delay its kicking in by 187.5ms (1000ms>>3 + 1000ms>>4).