re-add code for detecting dangling stress-marks
Code to detect dangling stress marks was removed which caused some
words to be incorrectly stressed. Also bring back the correct
check for german reverted in the previous commit, as the change in
this commit is the correct fix and the initial commit changing the
check for german was also correct (but broken by this missing code).