add parser rule compiler "par_c2" from tools/dtparser
This tool can be used to recompile the header par_rule2.h from the rules
in par_rule2.par (in dapi/src/cmd). It was imported from the tools-
archive from dtparser/par_c2 which is the newest version we have.

We haven't found a documentation of the rule-format yet, so writing
or modifying rules is currently not really feasible.

The current rules are responsible for the wrong expansion of Dr. and
St. in some cases (as they end the clause and there is no next word).