parser: special handling for period after phoneme
If a period directly follows a phoneme the period was spoken instead
of being used as a clause delimiter adding the period pause and forcing
a sync in the old version.

Add code to replicate the behaviour of the old version by replacing
the period with a period-phoneme directly followed by a sync-char
(0x0b, ctrl-k).

Closes #2