parser: handle non prefix-free R phoneme
Old versions seem to cope with R directly followed by another phoneme
even when the first char of the following phoneme is also a valid
phoneme when appended to R. The current code uses the "valid" two
character phoneme starting with R only to blow up shortly afterwards
when then next chars do not produce a valid phoneme (as the letter
after R is actually the first letter of the next phoneme).

This commit adds a workaround for this by analyzing the characters
following the R until a delimiter char (<:]) and if they are valid
phonemes, the input is used un-altered. If an error occurs, a space
is inserted after the R before the next characters. No further check
for invalid phonemes is done.

After this change I can't find any old songs still producing a
"command error in phoneme".