cm_pars: add a short breath break in front of sync
It seems old versions added a short breath break right in front of
a sync which is needed by the old intonation code. ttsapi.c seems
to have changed in this area quite a bit and it's no longer possible
to inject the breath break there, so re-add it in the parser instead.

Fixes unspoken [k] (and probably [p,t,b,d,g]) phoneme when using the
old intonation code. The new intonation code handles this differently
and doesn't need the breath break.