VTM1: run limit cycle fix earlier but not so aggressive
The limit cycle fix was cutting off /t/ phonemes at the end of
utterances previously, so it was modified to come into effect later.
This fixed the /t/ but also caused limit cycles (quiet beeps) to
reappear in the output for a short while.
Tackle this problem in another way by running the limit cycle fix
immediately but lessen the impact by a factor of 50. This clears
the signal from limit-cycles but also keeps the /t/ at the end
sounding good.