VTM1: use different constants for sixth parallel formant on 22kHz
The sixth parallel formant (responsible for /s/) uses hardcoded
constants with what seems a very high frequency range (seems to be
10760 Hz), which explains the heavy s-sounds on 22kHz as it is using
nearly the full possible range. Reduce this to the values found in
a comment in VTMF when on 22kHz, this makes /s/ much more bearable.
Use the original values on 11kHz.