build-system: make sure to create symlinks only once
There is a race-condition where to ln processes might be running
at the same time and one will fail with "file already exists" even
though both are using -f, as both processes checked exactly at the
and then one of them wins creating the file.
Fix this by only creating the symlink once and use it as a dependency
in the Makefile.