linux_audio: fix fragment time calculation
The code was trying to calculate the time it takes to play a fragment
by dividing by the number of bytes played per second. This leads to
an inaccuracy as the result is stored in an integer and will cause a
higher CPU load once a few hundres fragments are played as the
calculated deadline has always expired.
Fix this by storing the start-time and always counting only bytes so
the deadline can be recalculated without acummulating errors.
Also set the minimum sleep time to 1ms to avoid a busy loop when the
calculated deadline expired but there are no fragments are available.

This reduces the CPU load quite significantly by about 60%.