linux_audio: make it possible to use multiple output methods
Previously, when pulseaudio was enabled, then OSS on linux was disabled
and when compiled on macOS then pulseaudio was not used even if it was
installed. Change this by probing each available output and using the
first one which is working. This is needed for the upcoming native
ALSA implementation.