linux_audio: add direct ALSA output driver
This implements direct audio output to ALSA instead of using Pulseaudio.
This is useful for systems which have no pulseaudio/pipewire-daemon
running because of limited ressources or because they are headless
systems.
If DECtalk is compiled with Pulseaudio and ALSA support, then Pulseaudio
will be tried first. If this does not work, ALSA is attempted and if
this still does not work OSS is used (if on a platform which supports
OSS).