change VDF/VOICE_ROM selection and use VOICE ROM from 4.3
Previously there was an implicit default when no voice ROM or no
voice definition was selected. Change this so an explicit selection
has to be made and set the default for voice definition and voice
ROM to ones extracted from SPEAK43.EXE (DECtalk 4.3 Demo for Windows).
The previous voice ROM was from 1996 extracted from microdectalk,
the voice definition already was from DECtalk 4.3.

Fixes: "[:nb] no, no, no." (Thanks @jaybird110127)