VTM: add [:spf NNN] to specify samples per frame in percent
Currently the synthesizer and VTM use a frame-length of 6.4ms which
translates to 71 audio samples per frame att 11kHz. This is fine for
normal speaking rates but increasing the rate to high levels causes
the output to "mumble".

Add a way to changes the samples per frame with the new [:spf NNN]
command, where NNN is the percentage of samples in regards to the
default number (100 = default, lower = faster, higher = slower).
This changes the speed of the output regardless of the setting in
[:rate], so you might want to change both to get the perfect
output for your case.

Try: [:ra 240 :spf 50 :version speak]