Add support for 10kHz sample rate
This adds support for the DECtalk hardware sample rate of 10kHz. There
might be places where something has been missed so consider this as
work in progress.

As PC audio hardware is usually running at either a multiple of 11025 Hz
or a multiple of 48000 Hz, this will most likely get resampled by your
OS/driver/sound hardware with various degrees of success.