say (unix): add clause mode (-c)
This adds clause-mode to say to only speak after a clause delimiter
(like period or comma) was entered. Normally say always forces speech
on line-endings but that might not always be what is expected as a
pause is added if a clause is spans over multiple lines.
This can be activated for stdin mode (-) and file-in mode (-fi) with
the new switch -c.