VTM1: adjust zapping to old code, extend cosine_table
The old code uses slightly different ranges for decisions on when to
zap the input. This uses the ranges from the old implementation, which
is off by one.
In addition the old code handles zapping in d2pole_cf123 by shifting
the sample rate which could lead to invalid memory accesses in the
cosine_table for sample rates greater than 10kHz. Extend the cosine
table to also support 11025Hz (higher sample rates do not use the table
but use on-the-fly calculation of the cosine).