configure: use CC as linker on linux
Using ld directly on linux might cause problems when gcc emits
references to symbols in libgcc which is the subsequently not
linked because ld is used directly instead of gcc.
Change the linker command to use $(CC) and adapt the parameters.

This fixes build and runtime-failures on 32bit armhf linux (raspios).

Fixes #38