linux_audio: fix ALSA fragment calculation and decrease latency target
The virtual fragment calculation for ALSA was wrong which lead to
combinations of hardware-drivers and latency targets to cause hanging
audio output because ALSA was still waiting for more data but dectalk
calculated that no buffer space is available. Fix the calculation to
count partial "fragments" as available.

This also makes it possible to reduce the latency target back to 100ms
without causing hangs on a raspberry pi.